调整代码结构，将配置文件处理相关的部分单独提取
